# openapi/image_analysis.yml
openapi: "3.0.2"
info:
  title: PROM Documents APIs
  version: "1.0"
servers:
  - url: https://api.dev.onpromtronics.com/api/documents
paths:
...
  /instruments/query:
    post:
      tags:
        - Instruments
      summary: query a instrument
      description: Create a instrument
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentRequest"
      responses:
        "200":
          description: The action was successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badRequestResponse"
  /instruments:
    post:
      tags:
        - Instruments
      summary: Create a instrument
      description: Create a instrument
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentRequest"
      responses:
        "200":
          description: The resource was created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badRequestResponse"
    get:
      summary: Get the instrument list belong to the user
      tags:
        - Instruments
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentsListResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    badRequestResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: the resource you request is invalid
...
    InstrumentRequest:
      type: object
      properties:
        name:
          type: string
          format: string
        owner:
          type: string
          format: string
        inventoryNumber:
          type: string
          format: string
    InstrumentsResponse:
      type: object
      properties:
        statusCode:
          type: string
          format: number
          example: 200
        message:
          type: string
          format: number
          example: "request succeed"
        data:
          type: object
          properties:
            instrumentId:
              type: string
              format: uuid
            name:
              type: string
              description: name of the instrument
            owner:
              type: string
              description: owner of the instrument
            inventoryNumber:
              type: string
              description: inventory number of the instrument
            createdAt:
              type: string
              format: date-time
              description: creating datetime of the instrument
            lastUpdatedAt:
              type: string
              format: date-time
              description: updating datetime of the instrument
        totalCount:
          type: number
          example: 1
    InstrumentsListResponse:
      type: object
      properties:
        statusCode:
          type: string
          format: number
          example: 200
        data:
          type: array
          items:
            $ref: '#/components/schemas/InstrumentsData'
        totalCount:
          type: number
          example: 1
    InstrumentsData:
      type: object
      properties:
        instrumentId:
          type: string
          format: uuid
        name:
          type: string
          description: name of the instrument
        owner:
          type: string
          description: owner of the instrument
        inventoryNumber:
          type: string
          description: inventory number of the instrument
        createdAt:
          type: string
          format: date-time
          description: creating datetime of the instrument
        lastUpdatedAt:
          type: string
          format: date-time
          description: updating datetime of the instrument
        totalCount:
          type: number
          example: 1